Script that deletes all cached stan models
This can be used from an ipython repl as follows:

```
%run empty_stan_cache.py
```

It will delete all `.pkl` files in the directory
`data/cached_stan_models`. Usually this kind of thing isn't needed as the
`StanModel_cache` function I copied from the pystan docs is quite clever, but I
think it struggles to recognise when a file that another model includes has changed.