bug: fix incorrect cell cytoplasm/extracellular logic
I missed that the 'extracellular' volume was needed to find the correct
concentration of the GLCx species. This caused a number to be multiplied by 100
that shouldn't have been. Fixing this made the simulation a lot more
reasonable, though something must still be wrong as some concentrations still
go below zero.