feat: tidy compartment logic
It should now be possible to specify compartments other than cell
cytoplasm. The implementation is a bit bad as the Stan functions will need to
be recompiled in order to change the compartments. I couldn't easily think of a
better way though so I guess this is something to fix in future if/when it
becomes a problem.