Feat: save warmup iters, use fewer cores, lower initial stepsize
A few changes to the sampling module that I found made the model a bit more
reliable.

I think lowering the initial stepsize prevents the sampler getting stuck early
on, which was a problem I hit a few times.

I also think it is maybe best not to use all available cores - by default
cmdstanpy uses <number_of_available_cores> minus two, so I thought why not
stick with that for now.

Finally I've also made the sampler save warmup iterations (in the latest
cmdstanpy this doesn't break anything). This is handy for keeping track of how
the sampler is doing - if it's stuck after one iteration then there's a good
chance something is wrong!